Move node storage into its own class
Continuing the disassembly of K8sStorageClient into smaller classes,
move get_image_data to get image information from that class into a
new NodeStorage class, which is now the only Kubernetes storage
required by the ImageService.
